% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_library.R
\docType{package}
\name{plotmatrix-package}
\alias{plotmatrix-package}
\title{A package to make charts from matrices (similar to tile charts) in R.}
\description{
A Package to Visualize matrices
}
\details{
This package has been inspired from the waffle chart package by hrbrmstr
License: SIL OFL 1.1\cr
URL:
}
