% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{plotmatrix}
\alias{plotmatrix}
\title{Make charts with matrices}
\usage{
plotmatrix(matrix, title = "", subtitle = "", low_color = "#132B43",
  high_color = "#56B1F7", legend_position = "top", xlabel = "",
  ylabel = "", title_size = 8, subtitle_size = 7, height = 0.9,
  width = 0.9, color = "black", key_size = 1.5, colour = "black")
}
\arguments{
\item{matrix}{named data.frame or matrix}

\item{subtitle}{subtitle of the Plot}

\item{low_color}{is the color assigned to the lowest value}

\item{high_color}{is the color assigned to the highest value}

\item{legend_position}{position of legend}

\item{xlabel}{x label for the chart}

\item{ylabel}{ylabel for the chart}

\item{title_size}{size of the title text}

\item{subtitle_size}{size of the subtitle text}

\item{height}{height of each cell}

\item{width}{width of each cell}

\item{color}{color of the cell borders}

\item{key_size}{size of legend keys}

\item{colour}{colour of the labels}

\item{header}{header of the Plot}
}
\description{
Given a named matrix or a data frame, this function will return a ggplot object that
represents a tile chart with the values of the matrix. The library gives the user the
choice to change the color , title and subtitle and the legend position in the plot.
}
\details{
If a data frame is used, the column will be taken as the column of the matrix and the rows
will be taken as the rows of the matrix and the function will plot accordingly.

A matrix input is also allowed.

Chart title and x-axis and y-axis labels are optional.



plotmatrix is inspired by hrbrmstr's waffle package (@hrbrmstr)
}
\examples{
data_matrix  <- data.frame(a=c(1,0,0),b=c(0,1,0),c=c(0,0,1))
plotmatrix(data_matrix)

# Plotting matrix with high and low color
plotmatrix(data_matrix,low_color="#f6efb9",high_color="#d6efd9")












}
